backend/requirements.txt
fastapi
uvicorn[standard]
requests 
sqlalchemy[asyncio]
aiosqlite 
python-dotenv 